Adds back read access to the Slack integration
This is important because read access is required to fetch
the user's first name, last name, and email address from Slack.

Currently the only place this is done is in server/app.js:84.
Also, to alay any fears, the token that is returned nor the
refresh token are ever saved on the selector's end.